feat/#Issue14 implemented LIC 10  (#37)
* feat/#issue14 Added CMV 2,1, LIC 10

Implemented LIC 10 and tests for the function. The commit also contains
a function that can calculate the area for a triangel by 3 points.

Replaced numpoints to points.size in LIC10.
Numpoints were were replaced in master branch by points.size. The same
changes where made in LIC 10 for the program to be able to run.

Reformated the code.
The classes Decide.java and DecideTest.java has been reformated to a
correct format.